requirements.txt
langchain_openai 
langchain_core
python-dotenv
langchain_community
Flask
PyPDF2